@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/sn-app",
-  "version": "1.9.0",
+  "version": "1.10.0",
   "main": "dist/index.js",
   "files": [
     "dist",
@@ -31,51 +31,52 @@
   "devDependencies": {
     "@relative-ci/agent": "^1.2.0",
     "@types/autosuggest-highlight": "^3.1.0",
-    "@types/react": "^16.9.41",
+    "@types/react": "^16.9.44",
     "@types/react-autosuggest": "^9.3.14",
     "@types/react-dom": "^16.9.8",
     "@types/react-responsive": "^8.0.2",
     "@types/react-router": "^5.1.8",
     "@types/react-router-dom": "^5.1.2",
     "@types/react-virtualized": "^9.21.10",
-    "@types/uuid": "^8.0.0",
+    "@types/uuid": "^8.0.1",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^6.0.3",
-    "css-loader": "^3.6.0",
-    "cypress": "^4.11.0",
+    "css-loader": "^4.2.1",
+    "cypress": "^4.12.1",
     "cypress-xpath": "^1.4.0",
     "file-loader": "^6.0.0",
-    "fork-ts-checker-webpack-plugin": "^5.0.5",
+    "fork-ts-checker-webpack-plugin": "^5.0.14",
     "html-webpack-plugin": "^4.0.4",
     "mini-css-extract-plugin": "^0.9.0",
     "monaco-editor-webpack-plugin": "^1.9.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "style-loader": "^1.1.3",
-    "terser-webpack-plugin": "^3.0.6",
-    "ts-loader": "^7.0.5",
+    "terser-webpack-plugin": "^4.0.0",
+    "ts-loader": "^8.0.2",
     "url-loader": "^4.0.0",
-    "webpack": "^4.42.0",
+    "webpack": "^4.44.1",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.2"
+    "webpack-merge": "^5.1.1"
   },
   "dependencies": {
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.56",
     "@sensenet/authentication-oidc-react": "^1.0.2",
-    "@sensenet/client-core": "^3.1.0",
-    "@sensenet/client-utils": "^1.8.4",
-    "@sensenet/controls-react": "^3.3.9",
-    "@sensenet/default-content-types": "^2.2.0",
-    "@sensenet/document-viewer-react": "^2.2.0",
-    "@sensenet/hooks-react": "^1.2.0",
+    "@sensenet/client-core": "^3.2.0",
+    "@sensenet/client-utils": "^1.9.0",
+    "@sensenet/controls-react": "^3.4.0",
+    "@sensenet/default-content-types": "^2.2.1",
+    "@sensenet/document-viewer-react": "^2.2.1",
+    "@sensenet/hooks-react": "^1.3.0",
     "@sensenet/icons-react": "^1.2.19",
-    "@sensenet/list-controls-react": "^1.4.4",
-    "@sensenet/pickers-react": "^1.3.0",
+    "@sensenet/list-controls-react": "^1.4.5",
+    "@sensenet/pickers-react": "^1.3.1",
     "@sensenet/query": "^1.1.15",
-    "@sensenet/repository-events": "^1.4.16",
+    "@sensenet/repository-events": "^1.5.0",
     "autosuggest-highlight": "^3.1.1",
     "clsx": "^1.1.1",
     "filesize": "^6.1.0",
@@ -89,7 +90,7 @@
     "react-responsive": "^8.1.0",
     "react-virtualized": "^9.21.2",
     "semaphore-async-await": "^1.5.1",
-    "uuid": "^8.2.0"
+    "uuid": "^8.3.0"
   },
   "typings": "./dist/index.d.ts"
 }
