@@ -1,7 +1,7 @@
 {
   "private": true,
   "name": "sn-react-component-docs",
-  "version": "3.0.1",
+  "version": "3.0.2",
   "description": "UI components for building apps upon sensenet implemented in React",
   "main": "index.js",
   "scripts": {
@@ -28,21 +28,21 @@
   },
   "homepage": "https://sensenet.com",
   "dependencies": {
-    "@sensenet/client-core": "^3.1.0",
-    "@sensenet/client-utils": "^1.8.4",
-    "@sensenet/controls-react": "^3.3.9",
-    "@sensenet/default-content-types": "^2.2.0",
-    "@sensenet/document-viewer-react": "^2.2.0",
+    "@sensenet/client-core": "^3.2.0",
+    "@sensenet/client-utils": "^1.9.0",
+    "@sensenet/controls-react": "^3.4.0",
+    "@sensenet/default-content-types": "^2.2.1",
+    "@sensenet/document-viewer-react": "^2.2.1",
     "@sensenet/icons-react": "^1.2.19",
-    "@sensenet/list-controls-react": "^1.4.4",
-    "@sensenet/pickers-react": "^1.3.0",
+    "@sensenet/list-controls-react": "^1.4.5",
+    "@sensenet/pickers-react": "^1.3.1",
     "@sensenet/query": "^1.1.15",
-    "@sensenet/search-react": "^1.2.20",
+    "@sensenet/search-react": "^1.3.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.4",
+    "@babel/core": "^7.11.1",
     "@storybook/addon-a11y": "^5.3.19",
     "@storybook/addon-actions": "^5.3.19",
     "@storybook/addon-events": "^5.3.19",
@@ -57,15 +57,15 @@
     "@storybook/core": "^5.3.19",
     "@storybook/react": "^5.3.19",
     "@storybook/source-loader": "^5.3.19",
-    "@types/node": "^14.0.14",
-    "@types/react": "^16.9.41",
-    "@types/uuid": "^8.0.0",
+    "@types/node": "^14.0.27",
+    "@types/react": "^16.9.44",
+    "@types/uuid": "^8.0.1",
     "babel-loader": "^8.0.6",
     "prop-types": "^15.7.2",
     "react-docgen-typescript-loader": "^3.7.1",
-    "regenerator-runtime": "^0.13.3",
-    "ts-loader": "^7.0.5",
-    "uuid": "^8.2.0",
-    "webpack": "^4.42.0"
+    "regenerator-runtime": "^0.13.7",
+    "ts-loader": "^8.0.2",
+    "uuid": "^8.3.0",
+    "webpack": "^4.44.1"
   }
 }
