@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/search-react",
-  "version": "1.2.20",
+  "version": "1.3.0",
   "description": "Search component in react with sensenet",
   "main": "dist/index.js",
   "typings": "./dist/index.d.ts",
@@ -29,42 +29,42 @@
   "homepage": "https://sensenet.com",
   "dependencies": {
     "@material-ui/core": "^4.11.0",
-    "@sensenet/client-core": "^3.1.0",
-    "@sensenet/client-utils": "^1.8.4",
-    "@sensenet/control-mapper": "^2.1.5",
-    "@sensenet/default-content-types": "^2.2.0",
+    "@sensenet/client-core": "^3.2.0",
+    "@sensenet/client-utils": "^1.9.0",
+    "@sensenet/control-mapper": "^2.1.6",
+    "@sensenet/default-content-types": "^2.2.1",
     "@sensenet/icons-react": "^1.2.19",
     "@sensenet/query": "^1.1.15",
     "autosuggest-highlight": "^3.1.1",
     "react": "^16.13.0",
     "react-autosuggest": "^10.0.0",
     "react-dom": "^16.13.0",
-    "tslib": "^2.0.0"
+    "tslib": "^2.0.1"
   },
   "devDependencies": {
     "@types/autosuggest-highlight": "3.1.0",
     "@types/enzyme": "^3.10.5",
-    "@types/jest": "^26.0.3",
-    "@types/react": "^16.9.41",
+    "@types/jest": "^26.0.9",
+    "@types/react": "^16.9.44",
     "@types/react-autosuggest": "^9.3.14",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.2",
     "awesome-typescript-loader": "^5.2.1",
     "copy-webpack-plugin": "^6.0.3",
-    "css-loader": "^3.6.0",
+    "css-loader": "^4.2.1",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
     "enzyme-to-json": "^3.5.0",
     "file-loader": "^6.0.0",
-    "jest": "^26.1.0",
+    "jest": "^26.2.2",
     "postcss-flexbugs-fixes": "^4.2.0",
     "postcss-loader": "^3.0.0",
     "source-map-loader": "^1.0.1",
     "style-loader": "^1.1.3",
-    "ts-jest": "^26.1.1",
-    "typescript": "^3.9.5",
+    "ts-jest": "^26.1.4",
+    "typescript": "^3.9.7",
     "url-loader": "^4.0.0",
-    "webpack": "^4.42.0",
+    "webpack": "^4.44.1",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
