@@ -1,7 +1,7 @@
 import { debounce } from '@sensenet/client-utils'
 import { GenericContent, ReferenceFieldSetting } from '@sensenet/default-content-types'
 import { Query, QueryExpression, QueryOperators } from '@sensenet/query'
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import Autosuggest, {
   AutosuggestPropsSingleSection,
   GetSuggestionValue,
@@ -25,9 +25,10 @@ export interface ReferenceFieldProps<T> {
   defaultValueIdOrPath?: string | number
   fetchItems: (fetchQuery: Query<T>) => Promise<T[]>
   onQueryChange?: (key: string, query: Query<T>) => void
-  onChange?: (item: T) => void
+  onChange?: (item: T | null) => void
   renderSuggestion?: RenderSuggestion<T>
   autoSuggestProps?: Partial<AutosuggestPropsSingleSection<T>>
+  triggerClear?: number
 }
 
 /**
@@ -43,79 +44,80 @@ export interface ReferenceFieldState<T extends GenericContent> {
 /**
  * Reference field picker component
  */
-export class ReferenceField<T extends GenericContent = GenericContent> extends React.Component<
-  ReferenceFieldProps<T>,
-  ReferenceFieldState<T>
-> {
-  constructor(props: ReferenceField<T>['props']) {
-    super(props)
-    this.handleSelect = this.handleSelect.bind(this)
-    this.fetchItems = debounce(this.fetchItems.bind(this), 500)
-    this.onChange = this.onChange.bind(this)
+export function ReferenceField<T extends GenericContent>(props: ReferenceFieldProps<T>) {
+  const [inputValue, setInputValue] = useState('')
+  const [items, setItems] = useState<T[]>([])
+
+  const displayName = props.fieldSetting?.DisplayName ?? props.label
+  const description = props.fieldSetting?.Description || ''
+  const helperText = props.helperText ?? props.fieldSetting?.Description ?? ''
+  const inputProps: InputProps<T> = {
+    value: inputValue,
+    onChange: (ev) => onChange((ev.target as HTMLInputElement).value),
+    id: 'CommandBoxInput',
+  }
+
+  useEffect(() => {
+    ;(async () => {
+      if (props.defaultValueIdOrPath) {
+        const fetchedItems = await props.fetchItems(
+          new Query((q) =>
+            isNaN(props.defaultValueIdOrPath as number)
+              ? q.equals('Path', props.defaultValueIdOrPath)
+              : q.equals('Id', props.defaultValueIdOrPath),
+          ),
+        )
+        if (fetchedItems.length === 1 && fetchedItems[0]) {
+          const [item] = fetchedItems
+
+          setInputValue(item.DisplayName || item.Name)
+
+          props.onQueryChange?.(
+            props.fieldKey || props.fieldName.toString(),
+            new Query((q) => q.equals(props.fieldName, item.Id)),
+          )
+        }
+      }
+    })()
+  }, [])
+
+  useEffect(() => {
+    if (props.triggerClear) {
+      setInputValue('')
+    }
+  }, [props.triggerClear])
+
+  const onChange = (value: string) => {
+    setInputValue(value)
+    props.onChange?.(null)
+    fetchItems(getQueryFromTerm(`*${value}*`))
+  }
+
+  const fetchItems = debounce(async (query: Query<T>) => {
+    const fetchedItems = await props.fetchItems(query)
+    setItems(fetchedItems)
+  }, 500)
+
+  const getSuggestionValue: GetSuggestionValue<T> = (c) => {
+    return c.DisplayName || c.Name
+  }
+  const onSuggestionSelected: OnSuggestionSelected<T> = (_ev, data) => {
+    _ev.preventDefault()
+    handleSelect(data.suggestion)
   }
 
   /**
-   * Initial state
+   *
+   * @param term AutoComplete field value
    */
-  public state: ReferenceFieldState<T> = {
-    inputValue: '',
-    items: [],
-    selected: null,
-    renderSuggestion: (item, props) => (
-      <ReferenceFieldSuggestion item={item} query={props.query} isHighlighted={props.isHighlighted} />
-    ),
-  }
-
-  public onChange = (inputValue: string) => {
-    this.setState({ inputValue })
-    this.fetchItems(this.getQueryFromTerm(`*${inputValue}*`))
-  }
-
-  public async fetchItems(query: Query<T>) {
-    const items = await this.props.fetchItems(query)
-    this.setState({ items })
-  }
-
-  public getSuggestionValue: GetSuggestionValue<T> = (c) => {
-    return c.DisplayName || c.Name
-  }
-  public onSuggestionSelected: OnSuggestionSelected<T> = (_ev, data) => {
-    _ev.preventDefault()
-    this.handleSelect(data.suggestion)
-  }
-
-  public async componentDidMount() {
-    if (this.props.defaultValueIdOrPath) {
-      const items = await this.props.fetchItems(
-        new Query((q) =>
-          isNaN(this.props.defaultValueIdOrPath as number)
-            ? q.equals('Path', this.props.defaultValueIdOrPath)
-            : q.equals('Id', this.props.defaultValueIdOrPath),
-        ),
-      )
-      if (items.length === 1 && items[0]) {
-        const [item] = items
-        this.setState({
-          inputValue: item.DisplayName || item.Name,
-          selected: item,
-        })
-        this.props.onQueryChange &&
-          this.props.onQueryChange(
-            this.props.fieldKey || this.props.fieldName.toString(),
-            new Query((q) => q.equals(this.props.fieldName, item.Id)),
-          )
-      }
-    }
-  }
-
-  public getQueryFromTerm<TQueryReturns>(term: string) {
+  function getQueryFromTerm<TQueryReturns>(term: string) {
     const query = new Query<TQueryReturns>((q) =>
       q.query((q2) => q2.equals('Name', term).or.equals('DisplayName', term).or.equals('Path', term)),
     )
 
-    if (this.props.fieldSetting.AllowedTypes) {
+    if (props.fieldSetting.AllowedTypes) {
       new QueryOperators(query).and.query((q2) => {
-        ;(this.props.fieldSetting.AllowedTypes as string[]).forEach((allowedType, index, array) => {
+        ;(props.fieldSetting.AllowedTypes as string[]).forEach((allowedType, index, array) => {
           new QueryExpression(q2.queryRef).term(`TypeIs:${allowedType}`)
           if (index < array.length - 1) {
             return new QueryOperators(q2.queryRef).or
@@ -125,9 +127,9 @@ export class ReferenceField<T extends GenericContent = GenericContent> extends R
       })
     }
 
-    if (this.props.fieldSetting.SelectionRoots && this.props.fieldSetting.SelectionRoots.length) {
+    if (props.fieldSetting.SelectionRoots?.length) {
       new QueryOperators(query).and.query((q2) => {
-        ;(this.props.fieldSetting.SelectionRoots as string[]).forEach((root, index, array) => {
+        ;(props.fieldSetting.SelectionRoots as string[]).forEach((root, index, array) => {
           new QueryExpression(q2.queryRef).inTree(root)
           if (index < array.length - 1) {
             return new QueryOperators(q2.queryRef).or
@@ -139,65 +141,53 @@ export class ReferenceField<T extends GenericContent = GenericContent> extends R
     return query
   }
 
-  private handleSelect(item: T) {
-    this.props.onChange && this.props.onChange(item)
-    this.setState({
-      inputValue: item.DisplayName || item.Name,
-      selected: item,
-    })
-    this.props.onQueryChange &&
-      this.props.onQueryChange(
-        this.props.fieldKey || this.props.fieldName.toString(),
-        new Query((q) => q.equals(this.props.fieldName, item.Id)),
-      )
-  }
-
-  /**
-   * Renders the component
-   */
-  public render() {
-    const displayName = this.props.fieldSetting?.DisplayName ?? this.props.label
-    const description = this.props.fieldSetting?.Description || ''
-    const helperText = this.props.helperText ?? this.props.fieldSetting?.Description ?? ''
-    const inputProps: InputProps<T> = {
-      value: this.state.inputValue,
-      onChange: (ev) => this.onChange((ev.target as HTMLInputElement).value),
-      id: 'CommandBoxInput',
-    }
-
-    return (
-      <Autosuggest
-        {...this.props.autoSuggestProps}
-        theme={{
-          suggestionsList: {
-            listStyle: 'none',
-            margin: 0,
-            padding: 0,
-          },
-        }}
-        suggestions={this.state.items}
-        onSuggestionsFetchRequested={async (req) => {
-          const query = this.getQueryFromTerm<T>(`*${req.value}*`)
-          const items = await this.props.fetchItems(query)
-          this.setState({ items })
-        }}
-        onSuggestionsClearRequested={() => {
-          this.setState({ items: [] })
-        }}
-        getSuggestionValue={this.getSuggestionValue}
-        renderSuggestion={this.state.renderSuggestion}
-        inputProps={inputProps}
-        renderInputComponent={(props) => (
-          <ReferenceFieldInput
-            displayName={displayName}
-            description={description}
-            helperText={helperText}
-            inputProps={props}
-          />
-        )}
-        renderSuggestionsContainer={ReferenceFieldContainer}
-        onSuggestionSelected={this.onSuggestionSelected}
-      />
+  const handleSelect = (item: T) => {
+    props.onChange?.(item)
+    setInputValue(item.DisplayName || item.Name)
+    props.onQueryChange?.(
+      props.fieldKey || props.fieldName.toString(),
+      new Query((q) => q.equals(props.fieldName, item.Id)),
     )
   }
+
+  return (
+    <Autosuggest
+      {...props.autoSuggestProps}
+      theme={{
+        suggestionsList: {
+          listStyle: 'none',
+          margin: 0,
+          padding: 0,
+        },
+      }}
+      suggestions={items}
+      onSuggestionsFetchRequested={async (req) => {
+        const query = getQueryFromTerm<T>(`*${req.value}*`)
+        const fetchedItems = await props.fetchItems(query)
+        setItems(fetchedItems)
+      }}
+      onSuggestionsClearRequested={() => {
+        setItems([])
+      }}
+      getSuggestionValue={getSuggestionValue}
+      renderSuggestion={(item, suggestionProps) => (
+        <ReferenceFieldSuggestion
+          item={item}
+          query={suggestionProps.query}
+          isHighlighted={suggestionProps.isHighlighted}
+        />
+      )}
+      inputProps={inputProps}
+      renderInputComponent={(inputComponentProps) => (
+        <ReferenceFieldInput
+          displayName={displayName}
+          description={description}
+          helperText={helperText}
+          inputProps={inputComponentProps}
+        />
+      )}
+      renderSuggestionsContainer={ReferenceFieldContainer}
+      onSuggestionSelected={onSuggestionSelected}
+    />
+  )
 }
