@@ -5,6 +5,7 @@ import {
   CurrentChildrenContext,
   CurrentContentContext,
   LoadSettingsContext,
+  useLogger,
   useRepository,
 } from '@sensenet/hooks-react'
 import { VirtualCellProps, VirtualDefaultCell, VirtualizedTable } from '@sensenet/list-controls-react'
@@ -13,7 +14,7 @@ import clsx from 'clsx'
 import React, { useCallback, useContext, useEffect, useState } from 'react'
 import { ResponsiveContext, ResponsivePersonalSettings } from '../../context'
 import { globals, useGlobalStyles } from '../../globalStyles'
-import { useSelectionService } from '../../hooks'
+import { useLocalization, useSelectionService } from '../../hooks'
 import { ContentBreadcrumbs } from '../ContentBreadcrumbs'
 import { ContentContextMenu } from '../context-menu/content-context-menu'
 import { useDialog } from '../dialogs'
@@ -27,10 +28,12 @@ import {
   DescriptionField,
   DisplayNameComponent,
   EmailField,
+  EnabledField,
   IconField,
   LockedField,
   PhoneField,
   ReferenceField,
+  RolesField,
 } from '.'
 
 const useStyles = makeStyles(() => {
@@ -71,7 +74,7 @@ export interface ContentListProps {
 }
 
 export const isReferenceField = (fieldName: string, repo: Repository, schema = 'GenericContent') => {
-  const refWhiteList = ['AllowedChildTypes']
+  const refWhiteList = ['AllowedChildTypes', 'AllRoles']
   const setting = repo.schemas.getSchemaByName(schema).FieldSettings.find((f) => f.Name === fieldName)
   return refWhiteList.indexOf(fieldName) !== -1 || (setting && setting.Type === 'ReferenceFieldSetting') || false
 }
@@ -91,6 +94,8 @@ export const ContentList: React.FunctionComponent<ContentListProps> = (props) =>
   const classes = useStyles()
   const globalClasses = useGlobalStyles()
   const { openDialog } = useDialog()
+  const logger = useLogger('ContentList')
+  const localization = useLocalization()
   const [selected, setSelected] = useState<GenericContent[]>([])
   const [activeContent, setActiveContent] = useState<GenericContent>(children[0])
   const [isFocused, setIsFocused] = useState(true)
@@ -325,6 +330,49 @@ export const ContentList: React.FunctionComponent<ContentListProps> = (props) =>
             <ActionsField onOpen={(ev) => openContext(ev, fieldOptions.rowData)} />
           </ContextMenuWrapper>
         )
+      case 'Enabled':
+        if (fieldOptions.rowData[fieldOptions.dataKey] !== undefined) {
+          return (
+            <ContextMenuWrapper onContextMenu={(ev) => openContext(ev, fieldOptions.rowData)}>
+              <EnabledField
+                enabled={fieldOptions.rowData[fieldOptions.dataKey] as boolean}
+                description={fieldSettings.Description ?? ''}
+                onChange={async (value: boolean) => {
+                  try {
+                    await repo.patch({
+                      idOrPath: fieldOptions.rowData.Id,
+                      content: { [fieldOptions.dataKey]: value },
+                    })
+                  } catch (error) {
+                    logger.error({
+                      message: localization.contentList.errorContentModification,
+                      data: {
+                        relatedContent: fieldOptions.rowData,
+                        details: { error },
+                      },
+                    })
+                    return false
+                  }
+
+                  return true
+                }}
+              />
+            </ContextMenuWrapper>
+          )
+        }
+        return null
+      case 'AllRoles':
+        if (Array.isArray(fieldOptions.rowData[fieldOptions.dataKey])) {
+          return (
+            <ContextMenuWrapper onContextMenu={(ev) => openContext(ev, fieldOptions.rowData)}>
+              <RolesField
+                roles={fieldOptions.rowData[fieldOptions.dataKey] as GenericContent[]}
+                directRoles={fieldOptions.rowData.DirectRoles}
+              />
+            </ContextMenuWrapper>
+          )
+        }
+        return null
       default:
         break
     }
@@ -342,15 +390,16 @@ export const ContentList: React.FunctionComponent<ContentListProps> = (props) =>
     ) {
       const expectedContent = fieldOptions.rowData[fieldOptions.dataKey] as GenericContent
       if (
-        expectedContent &&
-        expectedContent.Id &&
-        expectedContent.Type &&
-        expectedContent.Name &&
-        expectedContent.Path
+        (expectedContent &&
+          expectedContent.Id &&
+          expectedContent.Type &&
+          expectedContent.Name &&
+          expectedContent.Path) ||
+        Array.isArray(expectedContent)
       ) {
         return (
           <ContextMenuWrapper onContextMenu={(ev) => openContext(ev, fieldOptions.rowData)}>
-            <ReferenceField content={expectedContent} />
+            <ReferenceField content={expectedContent} fieldName={fieldOptions.dataKey} parent={fieldOptions.rowData} />
           </ContextMenuWrapper>
         )
       }
