@@ -1,10 +1,10 @@
 import { ODataParams } from '@sensenet/client-core'
 import { GenericContent } from '@sensenet/default-content-types'
 import { useLogger, useRepository } from '@sensenet/hooks-react'
-import { Button, DialogActions, DialogContent, DialogTitle, TextField } from '@material-ui/core'
+import { Button, DialogActions, DialogContent, TextField } from '@material-ui/core'
 import React, { useState } from 'react'
 import { useLocalization } from '../../hooks'
-import { useDialog } from '.'
+import { DialogTitle, useDialog } from '.'
 
 export type CheckInProps = {
   content: GenericContent
