@@ -6,7 +6,6 @@ import { LinearProgress } from '@material-ui/core'
 import Button from '@material-ui/core/Button'
 import DialogActions from '@material-ui/core/DialogActions'
 import DialogContent from '@material-ui/core/DialogContent'
-import DialogTitle from '@material-ui/core/DialogTitle'
 import List from '@material-ui/core/List'
 import ListItem from '@material-ui/core/ListItem'
 import ListItemIcon from '@material-ui/core/ListItemIcon'
@@ -16,7 +15,7 @@ import React, { useEffect, useState } from 'react'
 import { useGlobalStyles } from '../../globalStyles'
 import { useLocalization, useSnRoute } from '../../hooks'
 import { Icon } from '../Icon'
-import { useDialog } from './dialog-provider'
+import { DialogTitle, useDialog } from '.'
 
 export interface CopyMoveDialogProps {
   currentParent: GenericContent
