@@ -2,5 +2,7 @@ export * from './abstract-logger'
 export * from './colors'
 export * from './console-logger'
 export * from './log-entries'
-export * from './logger'
 export * from './logger-collection'
+export * from './logger'
+export * from './notification-logger'
+export * from './notification-service'
